MultiChannelGrabber: further improvements

+++ Qitom/widgets/figureWidget.cpp
@@ -496,7 +496,7 @@ RetVal FigureWidget::liveImage(
 		{
 			QSharedPointer<ito::Param> paramFormat = getParamByInvoke(cam.data(), "pixelFormat", retval);
 			const char * pixelFormat = paramFormat->getVal<const char*>();
-            double min, max = 0.0; 
+            int min, max = 0; 
             bool ok = false;
             AddInAbstractGrabber::integerPixelFormatStringToMinMaxValue(pixelFormat, min, max, ok);
             if (ok)
